[Feat]불필요한 api 삭제 및 정리
